[fix] FCM 메서드 모듈 파일로 독립.
