+++ src/services/UserService.ts
@@ -4,9 +4,7 @@ import User from "../models/User";
 import errorGenerator from '../errors/errorGenerator';
 import statusCode from '../modules/statusCode';
 import { UserInfoDto } from '../interfaces/user/UserInfoDto';
-import { UserNoticeBaseDto } from '../interfaces/user/UserAlarmBaseDto';
 
-import * as firebase from "firebase-admin";
 
 const updateUser = async (userId: any, userUpdateDto: UserUpdateDto) => {
   try {
@@ -90,48 +88,6 @@ const getUserForProfileUpdate = async (userId: any) => {
 
 }
 
-const pushAlarmToUser = async (userNoticeBaseDto: UserNoticeBaseDto) => {
-  try {
-    const user = await User.findById(userNoticeBaseDto.userId);
-    if (!user) {
-      return null;
-    }
-
-    if (user.isLogOut === true) {
-      throw errorGenerator({
-        msg: '로그아웃한 유저입니다.',
-        statusCode: statusCode.BAD_REQUEST
-      })
-    }
-
-    let alarm = {
-        notification: {
-          title: pushMessageTemplate.title,
-          body: pushMessageTemplate.body,
-        },
-        token: user.fcmToken as string
-      };
-
-    firebase
-      .messaging()
-      .send(alarm)
-      .then(function (res) {
-        console.log("성공적으로 메시지 발송 완료: ", res)
-      })
-      .catch(function (err) {
-        console.log('다음 메시지를 보내는 데 에러 발생: ', err)
-        throw errorGenerator({
-          msg: 'FCM 메시지 오류.',
-          statusCode: statusCode.INTERNAL_SERVER_ERROR
-        })
-      })
-
-
-  } catch (error) {
-    console.log(error);
-    throw error;
-  }
-}
 
 export default {
   updateUser,
@@ -139,5 +95,4 @@ export default {
   findUserByKakao,
   deleteUser,
   getUserForProfileUpdate,
-  pushAlarmToUser,
 };
